Better account for selected GM switches when compensating asymmetry (#455)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/455

WHAT?

Do not reset offset and always decrease asymmetry counter for non selected GMs

WHY?

1 - Changing non selected GM port changes the state of the system, effectively negatively affecting convergence
2 - Decreasing asymmetry counter always prevents GM switches from triggering asymmetry compensation directly

Reviewed By: vvfedorenko

Differential Revision: D72640986

fbshipit-source-id: 380126e931a397d8e3f9854d2aacc7e46ca6c231

+++ ptp/sptp/client/client.go
@@ -111,7 +111,7 @@ type Client struct {
 	asymmetric bool
 
 	// Running counter of consecutive asymmetric results received
-	asymmetryCounter uint16
+	asymmetryCounter int
 }
 
 func (c *Client) incrementSequence() {
